+++ config/sites.json
@@ -1,23 +1,1 @@
-[
-  {
-    "id": "67969971-3095-4f90-8680-96ecf01893c4",
-    "name": "Google",
-    "url": "https://www.google.com/",
-    "link": "https://www.google.com/",
-    "icon": "https://upload.wikimedia.org/wikipedia/commons/thumb/5/53/Google_%22G%22_Logo.svg/200px-Google_%22G%22_Logo.svg.png"
-  },
-  {
-    "name": "Facebook",
-    "url": "https://www.facebook.com/",
-    "link": "https://www.facebook.com/",
-    "icon": "https://www.facebook.com/images/fb_icon_325x325.png",
-    "id": "e7842240-f363-4d74-ac80-57fed0b59cfd"
-  },
-  {
-    "name": "LinkedIn",
-    "url": "https://www.linkedin.com/",
-    "link": "https://www.linkedin.com/",
-    "icon": "https://static-exp1.licdn.com/sc/h/al2o9zrvru7aqj8e1x2rzsrca",
-    "id": "1dd37871-4d68-4b37-a3ae-06935e3a5f62"
-  }
-]
+[]
