+++ config/users.json
@@ -1,26 +1,1 @@
-[
-  {
-    "id": "1c0a728b-59e5-4f8e-9d29-60b2e9bfb0e3",
-    "username": "admin",
-    "hash": "93ded223c3e570e258a1bdfe8941c7cc7ef5e1d062d19728418e42e1857cde260351bd23048ce0b4df3980ea105a9fd5a36c90acbec06cf4b2af09227f1d7d51ab99d415cf8afd8930a40795f885fe9152f6697cc1fd4b73c34ee25efeb84f0df5440444df4156c1f120a49e438f3165ed6f98631dc500ec0504c74de85b531a3b6854d5413561c2b519b3db6a18fe45ff546ec81239ece8ddcde3a06b243b541cba9433317b2b3dcae57d8846e4d9d9fed74371b16bb8096a3872d62b7a3ffdb50eaad3db746a2a1d01764ac0586943d30774fab6bcbd6c4e84f51a51f932f3582bb49ea1c38f3c2b669e1861df876fe88d93cb802717e5e4e8fdce379794570af2f8a3562f92249c6607b6e9ca0de02bcb918749b86c9b57843f16b8377548906cda02beab04f7058d47dfe58ca358744d69fd4a9a6eb1ca836f933b366db786bea0925270e0c393fbf4e3908f2b02c969c23e7f7ba4c62e5c34814653862a0622697f3a2a9e469acc8572f861c3964a610049343e4ac4a4405876a142a4a9adb3c0366bb13c91819824e88714eaa5f0f5a8e492358330bb300798ea8fd1160264f04be5e0a008b60571171b49c67c141ae40868e5d28a0bf32ce4ff72826900b8f700d10f451594fe8ce910f3aa48c2ca0de67486047ab4b1f62a5526fccb80cae6465ee8c7d0a2f53d0a421291b01ae12b070b732c11d58f44487f1238fa",
-    "salt": "598f6adf1232f5604a7c7fa2d5f031e2"
-  },
-  {
-    "id": "ede299c0-c1ac-456e-992a-b349c1b96ca7",
-    "username": "admin2",
-    "hash": "b5702c67c705954f8fbd4d6bca696e9ff10bc4068e980443df4bb83216cf3c0d0b3712722bdbbae95360eaa5b381b87d3d6586155bd6ec6ea95a9e15dc1f2dbe5bc33f98b8522d685365e5b036ac71820b319c08f79e1588038afd825b4ba6ffc5a2a898918d33a9d1e0411ccfdc1d42a7e17398153596ddab607074478d1e2a9c744d590fac104d4d9e21e52503f1177b21c051aa9e5b5d7a7d15458b881dff09e09ce485dbd4a25d2d26fcb78599370b2f5bb08f9d7ce70f9f7c38000f081a0816356f55283e6a00e251d6841b1e8dedf1e1c652ff9de6a72e58f7bf1389cb9e1820975c8694b0aa9706cc52cbffb41658367c5427854ca65f3dc90876880bb02005bf9ee31bb20b8e0f94e23dc8e161b5161ad3e65aa14de14c8283f4c085c623b67ffa44d96c2d82a441253225345cd8418239fd273f494bccd21040000967e971c780177fad1483c8813af9ea07fdd6cad8bfc5b0a767e92edace0189a51c30715f74269b4e010eb7f84f81d2c53cebd81ac67df8f92b7a567ef6db322e63a6733a15c9b5c97a949df4a1a05cfeca13830b539c729eb310cccaead9fcadc4d393f39a8a9a389655f0483d3994af0208e4357bb4eaf216663729498906d9bc33c9b391e264104215bfe7855cbafbe1e42737672094fab089324febb909a669a4b84b539fc7bd9203ab3b763af426688611dfa8e1ba98a87e9aab11ca300b",
-    "salt": "107dee926678113f90233698d765295a"
-  },
-  {
-    "id": "0b054238-8280-475b-9fe4-283de5c1bfe1",
-    "username": "admin3",
-    "hash": "a01d1daf36f8a40386ad6974b1fc8e559689251d8bf44cbd76dd91f27eef191e447e421823a854f56e2a2418053c91da63d40a7d3bdd60acff2ffc9a07457c0cc5a10ad8b5215a8f789d61099713c36eb4933d4f7575f135b30efff51245a040eb0dd09b40b3f523999ca90971e498d982c4f1b39dec1656d75f023d0633a71e9c1a8460abbdac4b7253f85f9810359e92d0c374e7cf92de77077093ec20ad6713751f334643c42a9828917c612c948ff6f6a55d86ac5bd0bc2e891b914b3a416152b8f5ec57db677498af750a0fec3a2595a35d7759ac420efce7d5ca3690d1c6f82bfdc041426871036ac231ef142cbff9d8c5b9ff40f24ff72455743d10cf2b7216e4fad6ac45ab802565525b921922f3a2a791e8c2d89f659a6c9869dcad844f34c2eee680742e9e19c01576ec9b175a9b3fccad6d93bb205a1f01fd4861f5e4ad9f836211843494b604358124c166722a57c7324c603447f885ed2c32c192d8098a73245595c5f23ef2bc9507d72826a981632312b4f8a171a34246eaf0122325861f2c2feb98a980a6c52d7e053e295d9beba70aeb8987adac83a75dd62159b1d972a3fee559fe6dd7a7d38b44154f99ab2d0334d4ecd4ed71a73accabf9c47bf6b84ab0f8743c238aa1bdb8bd8ced7bec8b7fdaf4f4f703241e7d677477c3d651cd07d7cabd1cd9e3fcb1020a5df45371046904b3a93d21d4dcc2f38e",
-    "salt": "8d410e79b7095359e90eb7a5d289c66c"
-  },
-  {
-    "id": "a2e88800-2fdf-49c7-a42f-01a91e8a937d",
-    "username": "admin4",
-    "hash": "423ae24719acd58fe34e9c4e8a30420df252927bbb5480947dbd38a68940a9de8ab4d7b67dee8f2388592bc3ac472fcd29afd13921c857163df2ad573f5330fd8b1c87bec61fc19d71252aa0206fae8199a94b4b9697facebb4fd4239ad2b0184bf1b73bb310148771330f875649ac91158a0c87b9ca83e0d742a30a662f93f0cb36e93507ed30b3457f2f805b1e68a560dfc3bbb2c7d9e20ba6e153daee894c34ba402f44d76c3aa79928cda66e86949189c3018b9ff01a8121578044354c099aa938007cbb9847a7bdf2c1141b19d4f7d57a8d5e3cd2c225fefcad52bee8bf26500623f7c1f087b7362d2e8b8837c1e2b2e05acc4896a7689dcf2068346fe78a0e6e0bb6885fefb275188a91e0d758b1ead9c83932c0c66a133739c789f133c5808a29a06aa9e059468c7d3c9a598f35d194d5f50f7ebc2fd95aab4bd9e8404340bd026eb14ac4f37eaf8fa8833c0ab61ec96825b0c4790cbc80944658c2fade7f629924a213cf637919a577fe78feab8fe20f34e43ab6783088e84e13ad9bab0b955d296ae465fed0a4b4f7e05a8dd9fe19c4bb49f70160343d60d46f2e1473be9896b919af88a3ffe7a428b4bbf422aa3636b6b8e22be0326abb92249b2b044285b85e2b93615f3204d13a1aef53981071788f075d170f5a72ae911e9b86e478ae172d2650f1329511407c831a9fd4e7aacfbbeabbed2d4333e2f128bda8",
-    "salt": "3ded2aa3146380871ea78276a1e93afe"
-  }
-]
+[]
