+++ routes/index.js
@@ -1,0 +1,3 @@
+import ping from './ping.route.js';
+
+export default { ping };
