--- routes/ping.js
@@ -1,26 +1,0 @@
-import { Router } from 'express';
-const router = Router();
-
-router.get('/', (req, res) => {
-  // Get all
-});
-
-router.post('/', (req, res) => {
-
-});
-
-router.get('/:id', (req, res) => {
-  // get data
-});
-
-router.post('/:id', (req, res) => {
-  // Initiate check now
-});
-
-router.put('/:id', (req, res) => {
-  // Update id
-});
-
-router.delete('/:id', (req, res) => {
-  // Delete id
-});
